Fixed sdl formula in upstream brew
To compile this repository in macOS, latest SDL 1.2 code is needed. Until now, `brew` formula to install `sdl` latest `HEAD` was broken. The installation line in 'readme' was using the formula file I proposed in a pull request to `homebrew-core` (https://github.com/Homebrew/homebrew-core/pull/54630). But now it is merged, so better to use the standard command `brew install sdl --HEAD`.